Update Crate and Implement Compression (#3)
* Format files with default settings

* Fix clippy warnings as of rustc 1.53

* Switch mmap provider to `memmap2`

The previous `memmap` crate is dormant and is not receiving
updates.

* Fix bzip2 compression string

As documented here: http://wiki.ros.org/Bags/Format/2.0#Chunk

The compression value corresponding to bzip2 is `bz2` not `bzip2`.

* Implemented bzip2 chunk decompression

This is a relatively implementation, simplying calling out to the bzip2
crate. It doesn't do anything clever to avoid allocation or implement
a zero-copy scheme.

* Implemented lz4 chunk decompression

This is a relatively implementation, simplying calling out to the lz4
crate. It doesn't do anything clever to avoid allocation or implement
a zero-copy scheme.

Co-authored-by: Paul Jakob Schroeder <paul.schroeder@tangramvision.com>